requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pip==23.3.1
python-dotenv==0.21.0
setuptools==68.2.2
wheel==0.41.2
yarl==1.9.4
Click==8.0.1
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
